refactor(decoder): move the serve stream mapper into the module
Move DiffusionStreamMapper, WireDelta, and their unit tests out of
server/wire.rs into decoder/serve.rs, next to the chat StreamDecoder on the
shared primitives. server/wire.rs keeps only the OpenAI SSE envelope framing
and consumes WireDelta from crate::decoder. Shared test fixtures (FakeDecoder,
step) move to decoder/testutil.rs so the two decoder test files stop carrying
their own copies.

Comments in the moved code reworded to positive form, free of em dashes and
prose semicolons.
